split facts and assumptions; move entries to separate file
